fix: listen for postEvents globally
